@@ -1,10 +1,7 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import router from './router'
+import "bootstrap/dist/css/bootstrap.min.css"
+import "bootstrap"
+import './index.css'
 
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).mount('#app')
