@@ -1,36 +1,38 @@
 <template>
-  <div id="app">
-    <alert-banner/>
-    <navbar></navbar>
-    <div class="site">
-      <router-view />
+    <div class="inner-app">
+        <Navbar />
+        <Hero />
+        <div class="site">
+        </div>
+        <Footer />
     </div>
-    <end></end>
-  </div>
 </template>
 
 <script>
 import Navbar from "./components/Navbar.vue";
-import End from "@/components/End.vue";
-import AlertBanner from './components/AlertBanner.vue';
+import Hero from "./components/Hero.vue";
+import Footer from "./components/Footer.vue";
 
 export default {
-  name: "app",
-  components: {
-    Navbar,
-    End,
-    AlertBanner,
-  },
+    name: "App",
+    components: {
+        Navbar,
+        Hero,
+        Footer,
+    },
 };
 </script>
 
 <style scoped>
-#app {
-  display: flex;
-  min-height: 100vh;
-  flex-direction: column;
+.inner-app {
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
 }
 .site {
-  flex: 1;
+    flex: 1;
 }
+/* * {
+    border: 1px solid red !important;
+} */
 </style>
